--- tests/test_Alertmanager_skipped.json
@@ -1,22 +1,0 @@
-{
-    "labels": {
-        "alertname": "Watchdog",
-        "prometheus": "monitoring/monitoring-prometheus-oper-prometheus",
-        "severity": "none"
-    },
-    "annotations": {
-        "message": "This is an alert meant to ensure that the entire alerting pipeline is functional.\nThis alert is always firing, therefore it should always be firing in Alertmanager\nand always fire against a receiver. There are integrations with various notification\nmechanisms that send a notification when this alert is not firing. For example the\n\"DeadMansSnitch\" integration in PagerDuty.\n"
-    },
-    "startsAt": "1970-01-01T00:00:00.520032135Z",
-    "endsAt": "1970-01-01T00:00:00.520032135Z",
-    "generatorURL": "http://localhost",
-    "status": {
-        "state": "active",
-        "silencedBy": [],
-        "inhibitedBy": []
-    },
-    "receivers": [
-        "null"
-    ],
-    "fingerprint": "1b43ca4565de75d2"
-}
