--- tests/test_Alertmanager_suppressed.json
@@ -1,34 +1,0 @@
-{
-    "annotations": {
-        "message": "Network interface \"murpel\" showing errors on node-exporter monitoring/monitoring-prometheus-node-exporter-4711"
-    },
-    "endsAt": "1970-01-01T00:00:12.345Z",
-    "fingerprint": "0ef7c4bd7a504b8d",
-    "receivers": [
-        {
-            "name": "null"
-        }
-    ],
-    "startsAt": "1970-01-01T00:00:01.345Z",
-    "status": {
-        "inhibitedBy": [],
-        "silencedBy": [
-            "bb043288-42a0-4315-8bae-15cde1d7e239"
-        ],
-        "state": "suppressed"
-    },
-    "updatedAt": "1970-01-01T00:00:12.345Z",
-    "generatorURL": "http://localhost",
-    "labels": {
-        "alertname": "Error",
-        "device": "murpel",
-        "endpoint": "metrics",
-        "instance": "127.0.0.1:9100",
-        "job": "node-exporter",
-        "namespace": "monitoring",
-        "pod": "monitoring-prometheus-node-exporter-4711",
-        "prometheus": "monitoring/monitoring-prometheus-oper-prometheus",
-        "service": "monitoring-prometheus-node-exporter",
-        "severity": "warning"
-    }
-}
