--- tests/test_Alertmanager_warning.json
@@ -1,24 +1,0 @@
-{
-    "labels": {
-        "alertname": "TargetDown",
-        "job": "kubelet",
-        "prometheus": "monitoring/monitoring-prometheus-oper-prometheus",
-        "severity": "warning"
-    },
-    "annotations": {
-        "message": "66.6% of the kubelet targets are down."
-    },
-    "startsAt": "1970-01-01T00:00:00.520032135Z",
-    "endsAt": "1970-01-01T00:00:00.520032135Z",
-    "updatedAt": "1970-01-01T00:00:00.520032135Z",
-    "generatorURL": "http://localhost",
-    "status": {
-        "state": "active",
-        "silencedBy": [],
-        "inhibitedBy": []
-    },
-    "receivers": [
-        "null"
-    ],
-    "fingerprint": "7be970c6e97b95c9"
-}
